move community pages to nested url format and upgrade next
