+++ utils/magic.js
@@ -1,4 +1,4 @@
 import { Magic } from "@magic-sdk/admin";
 
 /* initiate Magic instance */
-export const magic = new Magic('pk_live_9FE80BF9C9AF42BB');
+export const magic = new Magic('pk_test_6B38245C25FB03BF');
